+++ react/ProfileRules.js
@@ -1,5 +1,8 @@
 import React, { Component } from 'react'
+import { injectIntl, intlShape } from 'react-intl'
 import PropTypes from 'prop-types'
+import moment from 'moment'
+import msk from 'msk'
 
 import defaultRules from './rules/default'
 
@@ -16,19 +19,23 @@ class ProfileRules extends Component {
   }
 
   updateRules() {
-    const { shouldUseIOFetching, fetch, country } = this.props
+    const { shouldUseIOFetching, fetch, country, intl } = this.props
 
     const rulePromise = shouldUseIOFetching
       ? import(`./rules/${country}`)
       : fetch(country)
-    return this.fetchRules(rulePromise)
+    return this.fetchRules(rulePromise, intl)
   }
 
-  async fetchRules(rulePromise) {
+  async fetchRules(rulePromise, intl) {
     try {
       const ruleData = await rulePromise
       const rules = ruleData.default || ruleData
+
+      prepareDateRules(rules, intl)
+
       this.setState({ rules })
+      
       return rules
     } catch (error) {
       const errorType = this.parseError(error)
@@ -38,6 +45,9 @@ class ProfileRules extends Component {
             `Couldn't load rules for country ${errorType}, using default rules instead.`,
           )
         }
+        
+        prepareDateRules(defaultRules, intl)
+
         this.setState({ rules: defaultRules })
         return defaultRules
       }
@@ -67,6 +77,8 @@ class ProfileRules extends Component {
 }
 
 ProfileRules.propTypes = {
+  /** React-intl utility */
+  intl: intlShape.isRequired,
   /** Components that will receive the rules */
   children: PropTypes.element.isRequired,
   /** The country whose rules will be fetched and applied */
@@ -77,4 +89,23 @@ ProfileRules.propTypes = {
   fetch: PropTypes.func,
 }
 
-export default ProfileRules
+export default injectIntl(ProfileRules)
+
+export function filterDateType(fields) {
+  return fields.filter((rule) => rule.type === 'date')
+}
+
+export function prepareDateRules(rules, intl) {
+
+  setDateRuleValidations(filterDateType(rules.personalFields), intl)
+  setDateRuleValidations(filterDateType(rules.businessFields), intl)
+}
+
+function setDateRuleValidations(rules, intl) {
+  rules && rules.map(rule => {
+    rule.mask =  value => msk.fit(value, '99/99/9999')
+    rule.validate = value => moment(value,'L',intl.locale.toLowerCase()).isValid()
+    rule.display = value => moment(value,[moment.ISO_8601, 'L'], intl.locale.toLowerCase()).utc().format('L')
+    rule.submit = value => moment(value,'L',intl.locale.toLowerCase()).utc().format()
+  })
+}