+++ react/__tests__/ProfileRules.test.js
@@ -1,84 +1,77 @@
-import React from 'react'
-import ProfileRules from '../ProfileRules'
-import { shallow } from 'enzyme'
-import brRules from '../rules/BRA'
+import { IntlProvider } from 'react-intl'
+
+import { prepareDateRules, filterDateType} from '../ProfileRules'
 import defaultRules from '../rules/default'
 
-describe('ProfileRules', () => {
-  it('should load the defined rules', async () => {
-    // Arrange
-    const instance = shallow(
-      <ProfileRules
-        country={'BRA'}
-        fetch={country => import('../rules/' + country)}
-      >
-        <h1>It works!</h1>
-      </ProfileRules>,
-    ).instance()
+const intlProvider = new IntlProvider({ locale: 'pt-br' }, {})
+const { intl } = intlProvider.getChildContext()
 
-    // Act
-    const rules = await instance.componentDidMount()
+describe('ProfileRules aux functions', () => {
+  it('should filter the correct rules', () => {
+    const rules = defaultRules
 
-    // Assert
-    expect(rules).toEqual(brRules)
+    const filteredFields = filterDateType(rules.personalFields)
+
+    expect(filteredFields.length).toBe(1)
+    expect(filteredFields[0].name).toBe('birthDate')
   })
 
-  it('should render its children', async () => {
-    // Arrange
-    const wrapper = shallow(
-      <ProfileRules
-        country={'BRA'}
-        fetch={country => import('../rules/' + country)}
-      >
-        <h1>It works!</h1>
-      </ProfileRules>,
-    )
-    const instance = wrapper.instance()
+  it('should set the functions to the rules of type date', () => {
+    const rules = defaultRules
 
-    // Act
-    await instance.componentDidMount()
-    wrapper.update()
+    const birthDate = filterDateType(rules.personalFields)[0]
 
-    // Assert
-    expect(wrapper.find('h1')).toHaveLength(1)
+    expect(birthDate.mask).toBeUndefined()
+    expect(birthDate.validate).toBeUndefined()
+    expect(birthDate.display).toBeUndefined()
+    expect(birthDate.submit).toBeUndefined()
+
+    prepareDateRules(rules, intl)
+
+    expect(birthDate.mask).toBeDefined()
+    expect(birthDate.validate).toBeDefined()
+    expect(birthDate.display).toBeDefined()
+    expect(birthDate.submit).toBeDefined()
   })
 
-  it('should inject rules as a prop to its children', async () => {
-    // Arrange
-    const wrapper = shallow(
-      <ProfileRules
-        country={'BRA'}
-        fetch={country => import('../rules/' + country)}
-      >
-        <h1>It works!</h1>
-      </ProfileRules>,
-    )
-    const instance = wrapper.instance()
+  it('should mask correctly', () => {
+    const rules = defaultRules
+    const birthDate = filterDateType(rules.personalFields)[0]
 
-    // Act
-    await instance.componentDidMount()
-    wrapper.update()
+    prepareDateRules(rules, intl)
 
-    // Assert
-    expect(wrapper.find('h1').props().rules).toBe(brRules)
+    expect(birthDate.mask('100')).toBe('10/0')
+    expect(birthDate.mask('10/09')).toBe('10/09')
+    expect(birthDate.mask('10/091994')).toBe('10/09/1994')
   })
 
-  it('should provide default rules when country is unrecognized', async () => {
-    // Arrange
-    global.console = { warn: jest.fn() }
-    const instance = shallow(
-      <ProfileRules
-        country={'XXX'}
-        fetch={country => import('../rules/' + country)}
-      >
-        <h1>It works!</h1>
-      </ProfileRules>,
-    ).instance()
+  it('should validate correctly', () => {
+    const rules = defaultRules
+    const birthDate = filterDateType(rules.personalFields)[0]
 
-    // Act
-    const rules = await instance.componentDidMount()
+    prepareDateRules(rules, intl)
 
-    // Assert
-    expect(rules).toEqual(defaultRules)
+    expect(birthDate.validate('100')).toBeFalsy()
+    expect(birthDate.validate('10/09/1994')).toBeTruthy()
+    expect(birthDate.validate('29/02/1000')).toBeFalsy()
+  })
+
+  it('should display correctly', () => {
+    const rules = defaultRules
+    const birthDate = filterDateType(rules.personalFields)[0]
+
+    prepareDateRules(rules, intl)
+
+    expect(birthDate.display('10/09/1994')).toBe('10/09/1994')
+    expect(birthDate.display('2018-11-25T00:00:00')).toBe('25/11/2018')
+  })
+
+  it('should submit correctly', () => {
+    const rules = defaultRules
+    const birthDate = filterDateType(rules.personalFields)[0]
+
+    prepareDateRules(rules, intl)
+
+    expect(birthDate.submit('25/11/2018')).toMatch(new RegExp('2018-11-25T([0-9])([0-9]):00:00Z'))
   })
 })