+++ react/rules/CHL.js
@@ -1,4 +1,5 @@
-import chile from '@vtex/phone/countries/CHL'
+import chile from '@vtex/phone/countries/CHL' // Used for initialization purposes, do not remove it!
+
 import { getPhoneFields } from '../modules/phone'
 import regexValidation from '../modules/regexValidation'
 
@@ -95,6 +96,7 @@ export default {
       name: 'birthDate',
       maxLength: 30,
       label: 'birthDate',
+      type: 'date'
     },
   ],
   businessFields: [