+++ react/rules/ECU.js
@@ -1,4 +1,5 @@
-import ecuador from '@vtex/phone/countries/ECU'
+import ecuador from '@vtex/phone/countries/ECU' // Used for initialization purposes, do not remove it!
+
 import { getPhoneFields } from '../modules/phone'
 import regexValidation from '../modules/regexValidation'
 
@@ -101,6 +102,7 @@ export default {
       name: 'birthDate',
       maxLength: 30,
       label: 'birthDate',
+      type: 'date'
     },
   ],
   businessFields: [